Changing href for the "Page load events" other than javascripttutorial.net
By reading the resource from the JavaScript Tutorial regarding Page Events I observed that it mentions
that the events DOMContentLoaded, load, unload, beforeunload are triggered on "document" object, a
statement that holds true only for the first event.

It also includes an HTML code that uses AddEventListener() as global function and thus is attached to
the global object, "window". As a beginner to web development,I had to search on stackoverflow to find
out on what object "AddEventListener()" is called when there is not a target element.